requirements.txt
argparse
google-generative-ai
python-dotenv